+++ seed/ts-express/basic-auth-environment-variables/index.d.ts
@@ -1,0 +1,3 @@
+export * as SeedBasicAuthEnvironmentVariables from "./api";
+export { register } from "./register";
+export { SeedBasicAuthEnvironmentVariablesError } from "./errors";