+++ seed/ts-sdk/basic-auth-environment-variables/src/errors/index.ts
@@ -1,0 +1,2 @@
+export { SeedBasicAuthEnvironmentVariablesError } from "./SeedBasicAuthEnvironmentVariablesError";
+export { SeedBasicAuthEnvironmentVariablesTimeoutError } from "./SeedBasicAuthEnvironmentVariablesTimeoutError";